+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 - rename `target-python-y` to `python-y`;
 - rename `target-bt-y` to `bpftrace-y`;
 #### Added
+- compiler/attribute: test `btf_decl_tag`;
 - btf: add `btf_dump.c`;
 - glibc: add limits sbudir;
 - compiler: add float subdir;